test(vm): consolidate TIP-2935 unit tests into integration suite
