+++ framework/src/test/java/org/tron/core/db/HistoryBlockHashIntegrationTest.java
@@ -11,6 +11,7 @@ import com.google.protobuf.ByteString;
 import java.lang.reflect.Field;
 import java.util.Arrays;
 import java.util.concurrent.atomic.AtomicReference;
+import org.bouncycastle.util.encoders.Hex;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
@@ -325,6 +326,148 @@ public class HistoryBlockHashIntegrationTest extends BaseTest {
     assertEquals(balance, after.getBalance());
     assertTrue(chainBaseManager.getCodeStore().has(addr));
     assertTrue(chainBaseManager.getContractStore().has(addr));
+  }
+
+  @Test
+  public void deployCreatesCodeContractAndAccount() {
+    HistoryBlockHashUtil.deploy(dbManager);
+
+    byte[] addr = HistoryBlockHashUtil.HISTORY_STORAGE_ADDRESS;
+
+    assertTrue(chainBaseManager.getCodeStore().has(addr));
+    CodeCapsule code = chainBaseManager.getCodeStore().get(addr);
+    assertNotNull(code);
+    assertArrayEquals(HistoryBlockHashUtil.HISTORY_STORAGE_CODE, code.getData());
+
+    ContractCapsule contract = chainBaseManager.getContractStore().get(addr);
+    assertNotNull(contract);
+    SmartContract proto = contract.getInstance();
+    assertEquals(HistoryBlockHashUtil.HISTORY_STORAGE_NAME, proto.getName());
+    assertArrayEquals(addr, proto.getContractAddress().toByteArray());
+    assertEquals("version must be 0", 0, proto.getVersion());
+    assertEquals(100L, proto.getConsumeUserResourcePercent());
+    assertArrayEquals("originAddress must be the EIP-2935 system caller",
+        HistoryBlockHashUtil.HISTORY_DEPLOYER_ADDRESS,
+        proto.getOriginAddress().toByteArray());
+
+    assertTrue(chainBaseManager.getAccountStore().has(addr));
+    AccountCapsule account = chainBaseManager.getAccountStore().get(addr);
+    assertEquals(HistoryBlockHashUtil.HISTORY_STORAGE_NAME,
+        account.getAccountName().toStringUtf8());
+    assertEquals(Protocol.AccountType.Contract, account.getType());
+    assertTrue("install marker must flip after a successful deploy",
+        chainBaseManager.getDynamicPropertiesStore().isBlockHashHistoryInstalled());
+  }
+
+  @Test
+  public void writeStoresParentHashAtCorrectSlot() {
+    HistoryBlockHashUtil.deploy(dbManager);
+
+    long blockNum = 100L;
+    byte[] parentHash = new byte[32];
+    Arrays.fill(parentHash, (byte) 0xab);
+
+    BlockCapsule block = new BlockCapsule(
+        blockNum,
+        Sha256Hash.wrap(parentHash),
+        System.currentTimeMillis(),
+        ByteString.copyFrom(new byte[21]));
+
+    HistoryBlockHashUtil.write(dbManager, block);
+
+    DataWord readBack = readSlot(99L);
+    assertNotNull(readBack);
+    assertArrayEquals(parentHash, readBack.getData());
+  }
+
+  @Test
+  public void writeUsesRingBufferModulo() {
+    HistoryBlockHashUtil.deploy(dbManager);
+
+    // (8192 - 1) % 8191 = 0
+    long blockNum = 8192L;
+    byte[] parentHash = new byte[32];
+    Arrays.fill(parentHash, (byte) 0xcd);
+
+    BlockCapsule block = new BlockCapsule(
+        blockNum,
+        Sha256Hash.wrap(parentHash),
+        System.currentTimeMillis(),
+        ByteString.copyFrom(new byte[21]));
+
+    HistoryBlockHashUtil.write(dbManager, block);
+
+    DataWord readBack = readSlot(0L);
+    assertNotNull(readBack);
+    assertArrayEquals(parentHash, readBack.getData());
+  }
+
+  @Test
+  public void beforeDeployNothingIsWritten() {
+    assertFalse(chainBaseManager.getCodeStore()
+        .has(HistoryBlockHashUtil.HISTORY_STORAGE_ADDRESS));
+    assertFalse(chainBaseManager.getContractStore()
+        .has(HistoryBlockHashUtil.HISTORY_STORAGE_ADDRESS));
+    assertFalse(chainBaseManager.getAccountStore()
+        .has(HistoryBlockHashUtil.HISTORY_STORAGE_ADDRESS));
+  }
+
+  /**
+   * If {@code deploy()} never ran (e.g. flag flipped without the deploy path),
+   * {@code write()} must not mutate {@code StorageRowStore} at the canonical
+   * address — otherwise the next call to {@code deploy()} would land on top of
+   * partially-written state.
+   */
+  @Test
+  public void writeIsNoOpBeforeDeploy() {
+    long blockNum = 100L;
+    byte[] parentHash = new byte[32];
+    Arrays.fill(parentHash, (byte) 0xab);
+    BlockCapsule block = new BlockCapsule(
+        blockNum,
+        Sha256Hash.wrap(parentHash),
+        System.currentTimeMillis(),
+        ByteString.copyFrom(new byte[21]));
+
+    HistoryBlockHashUtil.write(dbManager, block);
+
+    assertNull("write() must be a no-op without an installed BlockHashHistory",
+        readSlot(99L));
+  }
+
+  /**
+   * Defense-in-depth: when foreign bytecode sits at the canonical address,
+   * {@code deploy()} skips and the install marker stays 0, so {@code write()}
+   * must refuse to overwrite that contract's storage every block. Triggering
+   * the collision in practice requires a SHA-3 pre-image of the address, but
+   * the marker check is a single cached store hit.
+   */
+  @Test
+  public void writeIsNoOpOnForeignCode() {
+    byte[] addr = HistoryBlockHashUtil.HISTORY_STORAGE_ADDRESS;
+    byte[] foreignCode = Hex.decode("60016002");
+    chainBaseManager.getCodeStore().put(addr, new CodeCapsule(foreignCode));
+
+    HistoryBlockHashUtil.deploy(dbManager);
+
+    assertFalse("install marker must stay 0 when deploy skipped",
+        chainBaseManager.getDynamicPropertiesStore().isBlockHashHistoryInstalled());
+
+    long blockNum = 100L;
+    byte[] parentHash = new byte[32];
+    Arrays.fill(parentHash, (byte) 0xcd);
+    BlockCapsule block = new BlockCapsule(
+        blockNum,
+        Sha256Hash.wrap(parentHash),
+        System.currentTimeMillis(),
+        ByteString.copyFrom(new byte[21]));
+
+    HistoryBlockHashUtil.write(dbManager, block);
+
+    assertNull("write() must not overwrite a foreign contract's storage",
+        readSlot(99L));
+    assertArrayEquals("foreign code must remain intact",
+        foreignCode, chainBaseManager.getCodeStore().get(addr).getData());
   }
 
 }
